login/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework
djangorestframework-simplejwt
psycopg2-binary==2.9.10
sqlparse==0.5.3
django-environ==0.11.2
requests==2.31.0
